marking messages as sensitive in the helper_metadata instead of scrubbing contents in the template
